Abstract over pagination, make Github client closer to existing integrations
