--- integrations/github/port.py
@@ -1,9 +1,0 @@
-from enum import StrEnum
-
-
-class PortGithubResources(StrEnum):
-    REPO = "repository"
-    PR = "pull-requests"
-    ISSUE = "issues"
-    TEAM = "teams"
-    WORKFLOW = "workflows"
