--- integrations/github/wrappers/__init__.py
@@ -1,3 +1,0 @@
-from .github import GitHub
-
-__all__ = ["GitHub"]
