+++ integrations/github_melody/github/__init__.py
@@ -1,0 +1,3 @@
+from .client import GitHub
+
+__all__ = ["GitHub"]
