+++ integrations/github_melody/pyproject.toml
